refactor: remove ‘./index’ imports
