+++ src/utils/test/sort.test.ts
@@ -1,4 +1,4 @@
-import { sort } from '../index'
+import { sort } from '../sort'
 
 const target = {
   color: 'red',
